Crate bpf_sys

Modules

headers

perf_reader

uname

Structs

__BindgenBitfieldUnit

__IncompleteArrayField

__fsid_t

__kernel_fsid_t

__kernel_sockaddr_storage

__kernel_sockaddr_storage__bindgen_ty_1__bindgen_ty_1

__pthread_cond_s

__pthread_cond_s__bindgen_ty_1__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2__bindgen_ty_1

__pthread_internal_list

__pthread_mutex_s

__pthread_rwlock_arch_t

__sigset_t

__sk_buff

bpf_attr__bindgen_ty_1

bpf_attr__bindgen_ty_2

bpf_attr__bindgen_ty_3

bpf_attr__bindgen_ty_4

bpf_attr__bindgen_ty_5

bpf_attr__bindgen_ty_6

bpf_attr__bindgen_ty_7

bpf_attr__bindgen_ty_8

bpf_attr__bindgen_ty_9

bpf_attr__bindgen_ty_10

bpf_attr__bindgen_ty_11

bpf_attr__bindgen_ty_12

bpf_btf_info

bpf_cgroup_dev_ctx

bpf_cgroup_storage_key

bpf_create_map_attr

bpf_fib_lookup

bpf_flow_keys

bpf_flow_keys__bindgen_ty_1__bindgen_ty_1

bpf_flow_keys__bindgen_ty_1__bindgen_ty_2

bpf_func_info

bpf_insn

bpf_line_info

bpf_load_program_attr

bpf_lpm_trie_key

bpf_map_def

bpf_map_info

bpf_perf_event_value

bpf_prog_info

bpf_raw_tracepoint_args

bpf_sock

bpf_sock_addr

bpf_sock_ops

bpf_sock_tuple

bpf_sock_tuple__bindgen_ty_1__bindgen_ty_1

bpf_sock_tuple__bindgen_ty_1__bindgen_ty_2

bpf_sockopt

bpf_spin_lock

bpf_stack_build_id

bpf_sysctl

bpf_tcp_sock

bpf_tunnel_key

bpf_xdp_sock

bpf_xfrm_state

fd_set

if_stats_msg

ifla_bridge_id

ifla_cacheinfo

ifla_port_vsi

ifla_rmnet_flags

ifla_vf_broadcast

ifla_vf_guid

ifla_vf_link_state

ifla_vf_mac

ifla_vf_rate

ifla_vf_rss_query_en

ifla_vf_spoofchk

ifla_vf_trust

ifla_vf_tx_rate

ifla_vf_vlan

ifla_vf_vlan_info

ifla_vlan_flags

ifla_vlan_qos_mapping

ifla_vxlan_port_range

list_head

max_align_t

nl_mmap_hdr

nl_mmap_req

nl_pktinfo

nla_bitfield32

nlattr

nlmsgerr

nlmsghdr

rtnl_link_ifmap

rtnl_link_stats

rtnl_link_stats64

sk_msg_md

sk_reuseport_md

sockaddr_nl

timespec

timeval

xdp_md

Constants

BIG_ENDIAN

BPF_ABS

BPF_ADD

BPF_ADJ_ROOM_ENCAP_L2_MASK

BPF_ADJ_ROOM_ENCAP_L2_SHIFT

BPF_ALU

BPF_ALU64

BPF_AND

BPF_ANY

BPF_ARSH

BPF_B

BPF_BUILD_ID_SIZE

BPF_CALL

BPF_DEVCG_ACC_MKNOD

BPF_DEVCG_ACC_READ

BPF_DEVCG_ACC_WRITE

BPF_DEVCG_DEV_BLOCK

BPF_DEVCG_DEV_CHAR

BPF_DIV

BPF_DW

BPF_END

BPF_EXIST

BPF_EXIT

BPF_FIB_LKUP_RET_BLACKHOLE

BPF_FIB_LKUP_RET_FRAG_NEEDED

BPF_FIB_LKUP_RET_FWD_DISABLED

BPF_FIB_LKUP_RET_NOT_FWDED

BPF_FIB_LKUP_RET_NO_NEIGH

BPF_FIB_LKUP_RET_PROHIBIT

BPF_FIB_LKUP_RET_SUCCESS

BPF_FIB_LKUP_RET_UNREACHABLE

BPF_FIB_LKUP_RET_UNSUPP_LWT

BPF_FIB_LOOKUP_DIRECT

BPF_FIB_LOOKUP_OUTPUT

BPF_FLOW_DISSECTOR_F_PARSE_1ST_FRAG

BPF_FLOW_DISSECTOR_F_STOP_AT_ENCAP

BPF_FLOW_DISSECTOR_F_STOP_AT_FLOW_LABEL

BPF_FN_PREFIX

BPF_FROM_BE

BPF_FROM_LE

BPF_F_ADJ_ROOM_ENCAP_L4_GRE

BPF_F_ADJ_ROOM_ENCAP_L4_UDP

BPF_F_ADJ_ROOM_ENCAP_L3_IPV4

BPF_F_ADJ_ROOM_ENCAP_L3_IPV6

BPF_F_ADJ_ROOM_FIXED_GSO

BPF_F_ALLOW_MULTI

BPF_F_ALLOW_OVERRIDE

BPF_F_ANY_ALIGNMENT

BPF_F_CLONE

BPF_F_CTXLEN_MASK

BPF_F_CURRENT_CPU

BPF_F_CURRENT_NETNS

BPF_F_DONT_FRAGMENT

BPF_F_FAST_STACK_CMP

BPF_F_HDR_FIELD_MASK

BPF_F_INDEX_MASK

BPF_F_INGRESS

BPF_F_INVALIDATE_HASH

BPF_F_LOCK

BPF_F_MARK_ENFORCE

BPF_F_MARK_MANGLED_0

BPF_F_NO_COMMON_LRU

BPF_F_NO_PREALLOC

BPF_F_NUMA_NODE

BPF_F_PSEUDO_HDR

BPF_F_QUERY_EFFECTIVE

BPF_F_RDONLY

BPF_F_RDONLY_PROG

BPF_F_RECOMPUTE_CSUM

BPF_F_REUSE_STACKID

BPF_F_SEQ_NUMBER

BPF_F_SKIP_FIELD_MASK

BPF_F_STACK_BUILD_ID

BPF_F_STRICT_ALIGNMENT

BPF_F_SYSCTL_BASE_NAME

BPF_F_TEST_RND_HI32

BPF_F_TEST_STATE_FREQ

BPF_F_TUNINFO_IPV6

BPF_F_USER_BUILD_ID

BPF_F_USER_STACK

BPF_F_WRONLY

BPF_F_WRONLY_PROG

BPF_F_ZERO_CSUM_TX

BPF_F_ZERO_SEED

BPF_H

BPF_IMM

BPF_IND

BPF_JA

BPF_JEQ

BPF_JGE

BPF_JGT

BPF_JLE

BPF_JLT

BPF_JMP

BPF_JMP32

BPF_JNE

BPF_JSET

BPF_JSGE

BPF_JSGT

BPF_JSLE

BPF_JSLT

BPF_K

BPF_LD

BPF_LDX

BPF_LEN

BPF_LSH

BPF_MAXINSNS

BPF_MEM

BPF_MISC

BPF_MOD

BPF_MOV

BPF_MSH

BPF_MUL

BPF_NEG

BPF_NOEXIST

BPF_OBJ_NAME_LEN

BPF_OR

BPF_PSEUDO_CALL

BPF_PSEUDO_MAP_FD

BPF_PSEUDO_MAP_VALUE

BPF_REG_0

BPF_REG_1

BPF_REG_2

BPF_REG_3

BPF_REG_4

BPF_REG_5

BPF_REG_6

BPF_REG_7

BPF_REG_8

BPF_REG_9

BPF_REG_10

BPF_RET

BPF_RSH

BPF_SK_STORAGE_GET_F_CREATE

BPF_SOCK_OPS_ACTIVE_ESTABLISHED_CB

BPF_SOCK_OPS_ALL_CB_FLAGS

BPF_SOCK_OPS_BASE_RTT

BPF_SOCK_OPS_NEEDS_ECN

BPF_SOCK_OPS_PASSIVE_ESTABLISHED_CB

BPF_SOCK_OPS_RETRANS_CB

BPF_SOCK_OPS_RETRANS_CB_FLAG

BPF_SOCK_OPS_RTO_CB

BPF_SOCK_OPS_RTO_CB_FLAG

BPF_SOCK_OPS_RTT_CB

BPF_SOCK_OPS_RTT_CB_FLAG

BPF_SOCK_OPS_RWND_INIT

BPF_SOCK_OPS_STATE_CB

BPF_SOCK_OPS_STATE_CB_FLAG

BPF_SOCK_OPS_TCP_CONNECT_CB

BPF_SOCK_OPS_TCP_LISTEN_CB

BPF_SOCK_OPS_TIMEOUT_INIT

BPF_SOCK_OPS_VOID

BPF_ST

BPF_STX

BPF_SUB

BPF_TAG_SIZE

BPF_TCP_CLOSE

BPF_TCP_CLOSE_WAIT

BPF_TCP_CLOSING

BPF_TCP_ESTABLISHED

BPF_TCP_FIN_WAIT1

BPF_TCP_FIN_WAIT2

BPF_TCP_LAST_ACK

BPF_TCP_LISTEN

BPF_TCP_MAX_STATES

BPF_TCP_NEW_SYN_RECV

BPF_TCP_SYN_RECV

BPF_TCP_SYN_SENT

BPF_TCP_TIME_WAIT

BPF_TO_BE

BPF_TO_LE

BPF_W

BPF_X

BPF_XADD

BPF_XOR

BRIDGE_MODE_HAIRPIN

BRIDGE_MODE_UNSPEC

BYTE_ORDER

FD_SETSIZE

IFLA_ADDRESS

IFLA_AF_SPEC

IFLA_BOND_ACTIVE_SLAVE

IFLA_BOND_AD_ACTOR_SYSTEM

IFLA_BOND_AD_ACTOR_SYS_PRIO

IFLA_BOND_AD_INFO

IFLA_BOND_AD_INFO_ACTOR_KEY

IFLA_BOND_AD_INFO_AGGREGATOR

IFLA_BOND_AD_INFO_NUM_PORTS

IFLA_BOND_AD_INFO_PARTNER_KEY

IFLA_BOND_AD_INFO_PARTNER_MAC

IFLA_BOND_AD_INFO_UNSPEC

IFLA_BOND_AD_LACP_RATE

IFLA_BOND_AD_SELECT

IFLA_BOND_AD_USER_PORT_KEY

IFLA_BOND_ALL_SLAVES_ACTIVE

IFLA_BOND_ARP_ALL_TARGETS

IFLA_BOND_ARP_INTERVAL

IFLA_BOND_ARP_IP_TARGET

IFLA_BOND_ARP_VALIDATE

IFLA_BOND_DOWNDELAY

IFLA_BOND_FAIL_OVER_MAC

IFLA_BOND_LP_INTERVAL

IFLA_BOND_MIIMON

IFLA_BOND_MIN_LINKS

IFLA_BOND_MODE

IFLA_BOND_NUM_PEER_NOTIF

IFLA_BOND_PACKETS_PER_SLAVE

IFLA_BOND_PEER_NOTIF_DELAY

IFLA_BOND_PRIMARY

IFLA_BOND_PRIMARY_RESELECT

IFLA_BOND_RESEND_IGMP

IFLA_BOND_SLAVE_AD_ACTOR_OPER_PORT_STATE

IFLA_BOND_SLAVE_AD_AGGREGATOR_ID

IFLA_BOND_SLAVE_AD_PARTNER_OPER_PORT_STATE

IFLA_BOND_SLAVE_LINK_FAILURE_COUNT

IFLA_BOND_SLAVE_MII_STATUS

IFLA_BOND_SLAVE_PERM_HWADDR

IFLA_BOND_SLAVE_QUEUE_ID

IFLA_BOND_SLAVE_STATE

IFLA_BOND_SLAVE_UNSPEC

IFLA_BOND_TLB_DYNAMIC_LB

IFLA_BOND_UNSPEC

IFLA_BOND_UPDELAY

IFLA_BOND_USE_CARRIER

IFLA_BOND_XMIT_HASH_POLICY

IFLA_BROADCAST

IFLA_BRPORT_BACKUP_PORT

IFLA_BRPORT_BCAST_FLOOD

IFLA_BRPORT_BRIDGE_ID

IFLA_BRPORT_CONFIG_PENDING

IFLA_BRPORT_COST

IFLA_BRPORT_DESIGNATED_COST

IFLA_BRPORT_DESIGNATED_PORT

IFLA_BRPORT_FAST_LEAVE

IFLA_BRPORT_FLUSH

IFLA_BRPORT_FORWARD_DELAY_TIMER

IFLA_BRPORT_GROUP_FWD_MASK

IFLA_BRPORT_GUARD

IFLA_BRPORT_HOLD_TIMER

IFLA_BRPORT_ID

IFLA_BRPORT_ISOLATED

IFLA_BRPORT_LEARNING

IFLA_BRPORT_LEARNING_SYNC

IFLA_BRPORT_MCAST_FLOOD

IFLA_BRPORT_MCAST_TO_UCAST

IFLA_BRPORT_MESSAGE_AGE_TIMER

IFLA_BRPORT_MODE

IFLA_BRPORT_MULTICAST_ROUTER

IFLA_BRPORT_NEIGH_SUPPRESS

IFLA_BRPORT_NO

IFLA_BRPORT_PAD

IFLA_BRPORT_PRIORITY

IFLA_BRPORT_PROTECT

IFLA_BRPORT_PROXYARP

IFLA_BRPORT_PROXYARP_WIFI

IFLA_BRPORT_ROOT_ID

IFLA_BRPORT_STATE

IFLA_BRPORT_TOPOLOGY_CHANGE_ACK

IFLA_BRPORT_UNICAST_FLOOD

IFLA_BRPORT_UNSPEC

IFLA_BRPORT_VLAN_TUNNEL

IFLA_BR_AGEING_TIME

IFLA_BR_BRIDGE_ID

IFLA_BR_FDB_FLUSH

IFLA_BR_FORWARD_DELAY

IFLA_BR_GC_TIMER

IFLA_BR_GROUP_ADDR

IFLA_BR_GROUP_FWD_MASK

IFLA_BR_HELLO_TIME

IFLA_BR_HELLO_TIMER

IFLA_BR_MAX_AGE

IFLA_BR_MCAST_HASH_ELASTICITY

IFLA_BR_MCAST_HASH_MAX

IFLA_BR_MCAST_IGMP_VERSION

IFLA_BR_MCAST_LAST_MEMBER_CNT

IFLA_BR_MCAST_LAST_MEMBER_INTVL

IFLA_BR_MCAST_MEMBERSHIP_INTVL

IFLA_BR_MCAST_MLD_VERSION

IFLA_BR_MCAST_QUERIER

IFLA_BR_MCAST_QUERIER_INTVL

IFLA_BR_MCAST_QUERY_INTVL

IFLA_BR_MCAST_QUERY_RESPONSE_INTVL

IFLA_BR_MCAST_QUERY_USE_IFADDR

IFLA_BR_MCAST_ROUTER

IFLA_BR_MCAST_SNOOPING

IFLA_BR_MCAST_STARTUP_QUERY_CNT

IFLA_BR_MCAST_STARTUP_QUERY_INTVL

IFLA_BR_MCAST_STATS_ENABLED

IFLA_BR_MULTI_BOOLOPT

IFLA_BR_NF_CALL_ARPTABLES

IFLA_BR_NF_CALL_IP6TABLES

IFLA_BR_NF_CALL_IPTABLES

IFLA_BR_PAD

IFLA_BR_PRIORITY

IFLA_BR_ROOT_ID

IFLA_BR_ROOT_PATH_COST

IFLA_BR_ROOT_PORT

IFLA_BR_STP_STATE

IFLA_BR_TCN_TIMER

IFLA_BR_TOPOLOGY_CHANGE

IFLA_BR_TOPOLOGY_CHANGE_DETECTED

IFLA_BR_TOPOLOGY_CHANGE_TIMER

IFLA_BR_UNSPEC

IFLA_BR_VLAN_DEFAULT_PVID

IFLA_BR_VLAN_FILTERING

IFLA_BR_VLAN_PROTOCOL

IFLA_BR_VLAN_STATS_ENABLED

IFLA_BR_VLAN_STATS_PER_PORT

IFLA_CARRIER

IFLA_CARRIER_CHANGES

IFLA_CARRIER_DOWN_COUNT

IFLA_CARRIER_UP_COUNT

IFLA_COST

IFLA_EVENT

IFLA_EVENT_BONDING_FAILOVER

IFLA_EVENT_BONDING_OPTIONS

IFLA_EVENT_FEATURES

IFLA_EVENT_IGMP_RESEND

IFLA_EVENT_NONE

IFLA_EVENT_NOTIFY_PEERS

IFLA_EVENT_REBOOT

IFLA_EXT_MASK

IFLA_GENEVE_COLLECT_METADATA

IFLA_GENEVE_DF

IFLA_GENEVE_ID

IFLA_GENEVE_LABEL

IFLA_GENEVE_PORT

IFLA_GENEVE_REMOTE

IFLA_GENEVE_REMOTE6

IFLA_GENEVE_TOS

IFLA_GENEVE_TTL

IFLA_GENEVE_TTL_INHERIT

IFLA_GENEVE_UDP_CSUM

IFLA_GENEVE_UDP_ZERO_CSUM6_TX

IFLA_GENEVE_UDP_ZERO_CSUM6_RX

IFLA_GENEVE_UNSPEC

IFLA_GROUP

IFLA_GSO_MAX_SEGS

IFLA_GSO_MAX_SIZE

IFLA_GTP_FD0

IFLA_GTP_FD1

IFLA_GTP_PDP_HASHSIZE

IFLA_GTP_ROLE

IFLA_GTP_UNSPEC

IFLA_HSR_MULTICAST_SPEC

IFLA_HSR_SEQ_NR

IFLA_HSR_SLAVE1

IFLA_HSR_SLAVE2

IFLA_HSR_SUPERVISION_ADDR

IFLA_HSR_UNSPEC

IFLA_HSR_VERSION

IFLA_IFALIAS

IFLA_IFNAME

IFLA_IF_NETNSID

IFLA_INET6_UNSPEC

IFLA_INET6_FLAGS

IFLA_INET6_CONF

IFLA_INET6_STATS

IFLA_INET6_MCAST

IFLA_INET6_CACHEINFO

IFLA_INET6_TOKEN

IFLA_INET6_ADDR_GEN_MODE

IFLA_INET6_ICMP6STATS

IFLA_INET_CONF

IFLA_INET_UNSPEC

IFLA_INFO_DATA

IFLA_INFO_KIND

IFLA_INFO_SLAVE_DATA

IFLA_INFO_SLAVE_KIND

IFLA_INFO_UNSPEC

IFLA_INFO_XSTATS

IFLA_IPOIB_MODE

IFLA_IPOIB_PKEY

IFLA_IPOIB_UMCAST

IFLA_IPOIB_UNSPEC

IFLA_IPVLAN_FLAGS

IFLA_IPVLAN_MODE

IFLA_IPVLAN_UNSPEC

IFLA_LINK

IFLA_LINKINFO

IFLA_LINKMODE

IFLA_LINK_NETNSID

IFLA_MACSEC_CIPHER_SUITE

IFLA_MACSEC_ENCODING_SA

IFLA_MACSEC_ENCRYPT

IFLA_MACSEC_ES

IFLA_MACSEC_ICV_LEN

IFLA_MACSEC_INC_SCI

IFLA_MACSEC_PAD

IFLA_MACSEC_PORT

IFLA_MACSEC_PROTECT

IFLA_MACSEC_REPLAY_PROTECT

IFLA_MACSEC_SCB

IFLA_MACSEC_SCI

IFLA_MACSEC_UNSPEC

IFLA_MACSEC_VALIDATION

IFLA_MACSEC_WINDOW

IFLA_MACVLAN_FLAGS

IFLA_MACVLAN_MACADDR

IFLA_MACVLAN_MACADDR_COUNT

IFLA_MACVLAN_MACADDR_DATA

IFLA_MACVLAN_MACADDR_MODE

IFLA_MACVLAN_MODE

IFLA_MACVLAN_UNSPEC

IFLA_MAP

IFLA_MASTER

IFLA_MAX_MTU

IFLA_MIN_MTU

IFLA_MTU

IFLA_NET_NS_FD

IFLA_NET_NS_PID

IFLA_NEW_IFINDEX

IFLA_NEW_NETNSID

IFLA_NUM_RX_QUEUES

IFLA_NUM_TX_QUEUES

IFLA_NUM_VF

IFLA_OFFLOAD_XSTATS_CPU_HIT

IFLA_OFFLOAD_XSTATS_UNSPEC

IFLA_OPERSTATE

IFLA_PAD

IFLA_PHYS_PORT_ID

IFLA_PHYS_PORT_NAME

IFLA_PHYS_SWITCH_ID

IFLA_PORT_HOST_UUID

IFLA_PORT_INSTANCE_UUID

IFLA_PORT_PROFILE

IFLA_PORT_REQUEST

IFLA_PORT_RESPONSE

IFLA_PORT_SELF

IFLA_PORT_UNSPEC

IFLA_PORT_VF

IFLA_PORT_VSI_TYPE

IFLA_PPP_DEV_FD

IFLA_PPP_UNSPEC

IFLA_PRIORITY

IFLA_PROMISCUITY

IFLA_PROTINFO

IFLA_PROTO_DOWN

IFLA_QDISC

IFLA_RMNET_FLAGS

IFLA_RMNET_MUX_ID

IFLA_RMNET_UNSPEC

IFLA_STATS

IFLA_STATS64

IFLA_STATS_AF_SPEC

IFLA_STATS_LINK_64

IFLA_STATS_LINK_OFFLOAD_XSTATS

IFLA_STATS_LINK_XSTATS

IFLA_STATS_LINK_XSTATS_SLAVE

IFLA_STATS_UNSPEC

IFLA_TARGET_NETNSID

IFLA_TUN_GROUP

IFLA_TUN_MULTI_QUEUE

IFLA_TUN_NUM_DISABLED_QUEUES

IFLA_TUN_NUM_QUEUES

IFLA_TUN_OWNER

IFLA_TUN_PERSIST

IFLA_TUN_PI

IFLA_TUN_TYPE

IFLA_TUN_UNSPEC

IFLA_TUN_VNET_HDR

IFLA_TXQLEN

IFLA_UNSPEC

IFLA_VFINFO_LIST

IFLA_VF_BROADCAST

IFLA_VF_IB_NODE_GUID

IFLA_VF_IB_PORT_GUID

IFLA_VF_INFO

IFLA_VF_INFO_UNSPEC

IFLA_VF_LINK_STATE

IFLA_VF_LINK_STATE_AUTO

IFLA_VF_LINK_STATE_DISABLE

IFLA_VF_LINK_STATE_ENABLE

IFLA_VF_MAC

IFLA_VF_PORT

IFLA_VF_PORTS

IFLA_VF_PORT_UNSPEC

IFLA_VF_RATE

IFLA_VF_RSS_QUERY_EN

IFLA_VF_SPOOFCHK

IFLA_VF_STATS

IFLA_VF_STATS_BROADCAST

IFLA_VF_STATS_MULTICAST

IFLA_VF_STATS_PAD

IFLA_VF_STATS_RX_BYTES

IFLA_VF_STATS_RX_DROPPED

IFLA_VF_STATS_RX_PACKETS

IFLA_VF_STATS_TX_BYTES

IFLA_VF_STATS_TX_DROPPED

IFLA_VF_STATS_TX_PACKETS

IFLA_VF_TRUST

IFLA_VF_TX_RATE

IFLA_VF_UNSPEC

IFLA_VF_VLAN

IFLA_VF_VLAN_INFO

IFLA_VF_VLAN_INFO_UNSPEC

IFLA_VF_VLAN_LIST

IFLA_VLAN_EGRESS_QOS

IFLA_VLAN_FLAGS

IFLA_VLAN_ID

IFLA_VLAN_INGRESS_QOS

IFLA_VLAN_PROTOCOL

IFLA_VLAN_QOS_MAPPING

IFLA_VLAN_QOS_UNSPEC

IFLA_VLAN_UNSPEC

IFLA_VRF_PORT_TABLE

IFLA_VRF_PORT_UNSPEC

IFLA_VRF_TABLE

IFLA_VRF_UNSPEC

IFLA_VXLAN_AGEING

IFLA_VXLAN_COLLECT_METADATA

IFLA_VXLAN_DF

IFLA_VXLAN_GBP

IFLA_VXLAN_GPE

IFLA_VXLAN_GROUP

IFLA_VXLAN_GROUP6

IFLA_VXLAN_ID

IFLA_VXLAN_L2MISS

IFLA_VXLAN_L3MISS

IFLA_VXLAN_LABEL

IFLA_VXLAN_LEARNING

IFLA_VXLAN_LIMIT

IFLA_VXLAN_LINK

IFLA_VXLAN_LOCAL

IFLA_VXLAN_LOCAL6

IFLA_VXLAN_PORT

IFLA_VXLAN_PORT_RANGE

IFLA_VXLAN_PROXY

IFLA_VXLAN_REMCSUM_NOPARTIAL

IFLA_VXLAN_REMCSUM_RX

IFLA_VXLAN_REMCSUM_TX

IFLA_VXLAN_RSC

IFLA_VXLAN_TOS

IFLA_VXLAN_TTL

IFLA_VXLAN_TTL_INHERIT

IFLA_VXLAN_UDP_CSUM

IFLA_VXLAN_UDP_ZERO_CSUM6_TX

IFLA_VXLAN_UDP_ZERO_CSUM6_RX

IFLA_VXLAN_UNSPEC

IFLA_WEIGHT

IFLA_WIRELESS

IFLA_XDP

IFLA_XDP_ATTACHED

IFLA_XDP_DRV_PROG_ID

IFLA_XDP_FD

IFLA_XDP_FLAGS

IFLA_XDP_HW_PROG_ID

IFLA_XDP_PROG_ID

IFLA_XDP_SKB_PROG_ID

IFLA_XDP_UNSPEC

IFLA_XFRM_IF_ID

IFLA_XFRM_LINK

IFLA_XFRM_UNSPEC

INT8_MIN

INT8_MAX

INT16_MIN

INT16_MAX

INT32_MIN

INT32_MAX

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MIN

INT_FAST8_MAX

INT_FAST16_MIN

INT_FAST16_MAX

INT_FAST32_MIN

INT_FAST32_MAX

INT_LEAST8_MIN

INT_LEAST8_MAX

INT_LEAST16_MIN

INT_LEAST16_MAX

INT_LEAST32_MIN

INT_LEAST32_MAX

IPOIB_MODE_CONNECTED

IPOIB_MODE_DATAGRAM

IPVLAN_F_PRIVATE

IPVLAN_F_VEPA

LINK_XSTATS_TYPE_BOND

LINK_XSTATS_TYPE_BRIDGE

LINK_XSTATS_TYPE_UNSPEC

LITTLE_ENDIAN

LOG_BUF_SIZE

MACVLAN_FLAG_NOPROMISC

MAX_LINKS

MAX_VLAN_LIST_LEN

NETLINK_ADD_MEMBERSHIP

NETLINK_AUDIT

NETLINK_BROADCAST_ERROR

NETLINK_CAP_ACK

NETLINK_CONNECTED

NETLINK_CONNECTOR

NETLINK_CRYPTO

NETLINK_DNRTMSG

NETLINK_DROP_MEMBERSHIP

NETLINK_ECRYPTFS

NETLINK_EXT_ACK

NETLINK_FIB_LOOKUP

NETLINK_FIREWALL

NETLINK_GENERIC

NETLINK_GET_STRICT_CHK

NETLINK_INET_DIAG

NETLINK_IP6_FW

NETLINK_ISCSI

NETLINK_KOBJECT_UEVENT

NETLINK_LISTEN_ALL_NSID

NETLINK_LIST_MEMBERSHIPS

NETLINK_NETFILTER

NETLINK_NFLOG

NETLINK_NO_ENOBUFS

NETLINK_PKTINFO

NETLINK_RDMA

NETLINK_ROUTE

NETLINK_RX_RING

NETLINK_SCSITRANSPORT

NETLINK_SELINUX

NETLINK_SMC

NETLINK_SOCK_DIAG

NETLINK_TX_RING

NETLINK_UNCONNECTED

NETLINK_UNUSED

NETLINK_USERSOCK

NETLINK_XFRM

NET_MAJOR

NLA_ALIGNTO

NLA_F_NESTED

NLA_F_NET_BYTEORDER

NLA_TYPE_MASK

NLMSG_ALIGNTO

NLMSG_DONE

NLMSG_ERROR

NLMSG_MIN_TYPE

NLMSG_NOOP

NLMSG_OVERRUN

NLM_F_ACK

NLM_F_ACK_TLVS

NLM_F_APPEND

NLM_F_ATOMIC

NLM_F_CAPPED

NLM_F_CREATE

NLM_F_DUMP

NLM_F_DUMP_FILTERED

NLM_F_DUMP_INTR

NLM_F_ECHO

NLM_F_EXCL

NLM_F_MATCH

NLM_F_MULTI

NLM_F_NONREC

NLM_F_REPLACE

NLM_F_REQUEST

NLM_F_ROOT

NL_MMAP_MSG_ALIGNMENT

PDP_ENDIAN

PORT_PROFILE_MAX

PORT_PROFILE_RESPONSE_BADSTATE

PORT_PROFILE_RESPONSE_ERROR

PORT_PROFILE_RESPONSE_INPROGRESS

PORT_PROFILE_RESPONSE_INSUFFICIENT_RESOURCES

PORT_PROFILE_RESPONSE_INVALID

PORT_PROFILE_RESPONSE_SUCCESS

PORT_REQUEST_ASSOCIATE

PORT_REQUEST_DISASSOCIATE

PORT_REQUEST_PREASSOCIATE

PORT_REQUEST_PREASSOCIATE_RR

PORT_SELF_VF

PORT_UUID_MAX

PORT_VDP_RESPONSE_INSUFFICIENT_RESOURCES

PORT_VDP_RESPONSE_INVALID_FORMAT

PORT_VDP_RESPONSE_OUT_OF_SYNC

PORT_VDP_RESPONSE_SUCCESS

PORT_VDP_RESPONSE_UNUSED_VTID

PORT_VDP_RESPONSE_VTID_VERSION_VIOALTION

PORT_VDP_RESPONSE_VTID_VIOLATION

PTRDIFF_MAX

PTRDIFF_MIN

RMNET_FLAGS_EGRESS_MAP_CKSUMV4

RMNET_FLAGS_INGRESS_DEAGGREGATION

RMNET_FLAGS_INGRESS_MAP_CKSUMV4

RMNET_FLAGS_INGRESS_MAP_COMMANDS

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

TCP_BPF_IW

TCP_BPF_SNDCWND_CLAMP

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

XDP_ATTACHED_DRV

XDP_ATTACHED_HW

XDP_ATTACHED_MULTI

XDP_ATTACHED_NONE

XDP_ATTACHED_SKB

XDP_FLAGS_DRV_MODE

XDP_FLAGS_HW_MODE

XDP_FLAGS_MASK

XDP_FLAGS_MODES

XDP_FLAGS_SKB_MODE

XDP_FLAGS_UPDATE_IF_NOEXIST

XDP_PACKET_HEADROOM

_ATFILE_SOURCE

_BITS_BYTESWAP_H

_BITS_PTHREADTYPES_ARCH_H

_BITS_PTHREADTYPES_COMMON_H

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_UINTN_IDENTITY_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_ENDIAN_H

_FEATURES_H

_K_SS_MAXSIZE

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_STRUCT_TIMESPEC

_SYS_CDEFS_H

_SYS_SELECT_H

_SYS_TYPES_H

_THREAD_SHARED_TYPES_H

__BIG_ENDIAN

__BITS_PER_LONG

__BIT_TYPES_DEFINED__

__BYTE_ORDER

__FD_SETSIZE

__FD_ZERO_STOS

__FLOAT_WORD_ORDER

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__IFLA_BOND_AD_INFO_MAX

__IFLA_BOND_MAX

__IFLA_BOND_SLAVE_MAX

__IFLA_BRPORT_MAX

__IFLA_BR_MAX

__IFLA_GENEVE_MAX

__IFLA_GTP_MAX

__IFLA_HSR_MAX

__IFLA_INET6_MAX

__IFLA_INET_MAX

__IFLA_INFO_MAX

__IFLA_IPOIB_MAX

__IFLA_IPVLAN_MAX

__IFLA_MACSEC_MAX

__IFLA_MACVLAN_MAX

__IFLA_MAX

__IFLA_OFFLOAD_XSTATS_MAX

__IFLA_PORT_MAX

__IFLA_PPP_MAX

__IFLA_RMNET_MAX

__IFLA_STATS_MAX

__IFLA_TUN_MAX

__IFLA_VF_INFO_MAX

__IFLA_VF_LINK_STATE_MAX

__IFLA_VF_MAX

__IFLA_VF_PORT_MAX

__IFLA_VF_STATS_MAX

__IFLA_VF_VLAN_INFO_MAX

__IFLA_VLAN_MAX

__IFLA_VLAN_QOS_MAX

__IFLA_VRF_MAX

__IFLA_VRF_PORT_MAX

__IFLA_VXLAN_MAX

__IFLA_XDP_MAX

__IFLA_XFRM_MAX

__INO_T_MATCHES_INO64_T

__LINK_XSTATS_TYPE_MAX

__LITTLE_ENDIAN

__MAX_BPF_REG

__OFF_T_MATCHES_OFF64_T

__PDP_ENDIAN

__PTHREAD_MUTEX_HAVE_PREV

__PTHREAD_MUTEX_LOCK_ELISION

__PTHREAD_MUTEX_NUSERS_AFTER_KIND

__PTHREAD_MUTEX_USE_UNION

__PTHREAD_RWLOCK_INT_FLAGS_SHARED

__RLIM_T_MATCHES_RLIM64_T

__SIZEOF_PTHREAD_ATTR_T

__SIZEOF_PTHREAD_BARRIERATTR_T

__SIZEOF_PTHREAD_BARRIER_T

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__STDC_IEC_559__

__STDC_IEC_559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__clock_t_defined

__clockid_t_defined

__glibc_c99_flexarr_available

__have_pthread_attr_t

__sigset_t_defined

__time_t_defined

__timer_t_defined

__timeval_defined

bpf_adj_room_mode_BPF_ADJ_ROOM_MAC

bpf_adj_room_mode_BPF_ADJ_ROOM_NET

bpf_attach_type_BPF_CGROUP_DEVICE

bpf_attach_type_BPF_CGROUP_GETSOCKOPT

bpf_attach_type_BPF_CGROUP_INET4_BIND

bpf_attach_type_BPF_CGROUP_INET4_CONNECT

bpf_attach_type_BPF_CGROUP_INET4_POST_BIND

bpf_attach_type_BPF_CGROUP_INET6_BIND

bpf_attach_type_BPF_CGROUP_INET6_CONNECT

bpf_attach_type_BPF_CGROUP_INET6_POST_BIND

bpf_attach_type_BPF_CGROUP_INET_EGRESS

bpf_attach_type_BPF_CGROUP_INET_INGRESS

bpf_attach_type_BPF_CGROUP_INET_SOCK_CREATE

bpf_attach_type_BPF_CGROUP_SETSOCKOPT

bpf_attach_type_BPF_CGROUP_SOCK_OPS

bpf_attach_type_BPF_CGROUP_SYSCTL

bpf_attach_type_BPF_CGROUP_UDP4_SENDMSG

bpf_attach_type_BPF_CGROUP_UDP4_RECVMSG

bpf_attach_type_BPF_CGROUP_UDP6_SENDMSG

bpf_attach_type_BPF_CGROUP_UDP6_RECVMSG

bpf_attach_type_BPF_FLOW_DISSECTOR

bpf_attach_type_BPF_LIRC_MODE2

bpf_attach_type_BPF_SK_MSG_VERDICT

bpf_attach_type_BPF_SK_SKB_STREAM_PARSER

bpf_attach_type_BPF_SK_SKB_STREAM_VERDICT

bpf_attach_type___MAX_BPF_ATTACH_TYPE

bpf_cmd_BPF_BTF_GET_FD_BY_ID

bpf_cmd_BPF_BTF_GET_NEXT_ID

bpf_cmd_BPF_BTF_LOAD

bpf_cmd_BPF_MAP_CREATE

bpf_cmd_BPF_MAP_DELETE_ELEM

bpf_cmd_BPF_MAP_FREEZE

bpf_cmd_BPF_MAP_GET_FD_BY_ID

bpf_cmd_BPF_MAP_GET_NEXT_ID

bpf_cmd_BPF_MAP_GET_NEXT_KEY

bpf_cmd_BPF_MAP_LOOKUP_AND_DELETE_ELEM

bpf_cmd_BPF_MAP_LOOKUP_ELEM

bpf_cmd_BPF_MAP_UPDATE_ELEM

bpf_cmd_BPF_OBJ_GET

bpf_cmd_BPF_OBJ_GET_INFO_BY_FD

bpf_cmd_BPF_OBJ_PIN

bpf_cmd_BPF_PROG_ATTACH

bpf_cmd_BPF_PROG_DETACH

bpf_cmd_BPF_PROG_GET_FD_BY_ID

bpf_cmd_BPF_PROG_GET_NEXT_ID

bpf_cmd_BPF_PROG_LOAD

bpf_cmd_BPF_PROG_QUERY

bpf_cmd_BPF_PROG_TEST_RUN

bpf_cmd_BPF_RAW_TRACEPOINT_OPEN

bpf_cmd_BPF_TASK_FD_QUERY

bpf_func_id_BPF_FUNC_bind

bpf_func_id_BPF_FUNC_clone_redirect

bpf_func_id_BPF_FUNC_csum_diff

bpf_func_id_BPF_FUNC_csum_update

bpf_func_id_BPF_FUNC_current_task_under_cgroup

bpf_func_id_BPF_FUNC_fib_lookup

bpf_func_id_BPF_FUNC_get_cgroup_classid

bpf_func_id_BPF_FUNC_get_current_cgroup_id

bpf_func_id_BPF_FUNC_get_current_comm

bpf_func_id_BPF_FUNC_get_current_pid_tgid

bpf_func_id_BPF_FUNC_get_current_task

bpf_func_id_BPF_FUNC_get_current_uid_gid

bpf_func_id_BPF_FUNC_get_hash_recalc

bpf_func_id_BPF_FUNC_get_listener_sock

bpf_func_id_BPF_FUNC_get_local_storage

bpf_func_id_BPF_FUNC_get_numa_node_id

bpf_func_id_BPF_FUNC_get_prandom_u32

bpf_func_id_BPF_FUNC_get_route_realm

bpf_func_id_BPF_FUNC_get_smp_processor_id

bpf_func_id_BPF_FUNC_get_socket_cookie

bpf_func_id_BPF_FUNC_get_socket_uid

bpf_func_id_BPF_FUNC_get_stack

bpf_func_id_BPF_FUNC_get_stackid

bpf_func_id_BPF_FUNC_getsockopt

bpf_func_id_BPF_FUNC_ktime_get_ns

bpf_func_id_BPF_FUNC_l3_csum_replace

bpf_func_id_BPF_FUNC_l4_csum_replace

bpf_func_id_BPF_FUNC_lwt_push_encap

bpf_func_id_BPF_FUNC_lwt_seg6_store_bytes

bpf_func_id_BPF_FUNC_lwt_seg6_adjust_srh

bpf_func_id_BPF_FUNC_lwt_seg6_action

bpf_func_id_BPF_FUNC_map_delete_elem

bpf_func_id_BPF_FUNC_map_lookup_elem

bpf_func_id_BPF_FUNC_map_peek_elem

bpf_func_id_BPF_FUNC_map_pop_elem

bpf_func_id_BPF_FUNC_map_push_elem

bpf_func_id_BPF_FUNC_map_update_elem

bpf_func_id_BPF_FUNC_msg_apply_bytes

bpf_func_id_BPF_FUNC_msg_cork_bytes

bpf_func_id_BPF_FUNC_msg_pop_data

bpf_func_id_BPF_FUNC_msg_pull_data

bpf_func_id_BPF_FUNC_msg_push_data

bpf_func_id_BPF_FUNC_msg_redirect_hash

bpf_func_id_BPF_FUNC_msg_redirect_map

bpf_func_id_BPF_FUNC_override_return

bpf_func_id_BPF_FUNC_perf_event_output

bpf_func_id_BPF_FUNC_perf_event_read

bpf_func_id_BPF_FUNC_perf_event_read_value

bpf_func_id_BPF_FUNC_perf_prog_read_value

bpf_func_id_BPF_FUNC_probe_read

bpf_func_id_BPF_FUNC_probe_read_str

bpf_func_id_BPF_FUNC_probe_write_user

bpf_func_id_BPF_FUNC_rc_keydown

bpf_func_id_BPF_FUNC_rc_pointer_rel

bpf_func_id_BPF_FUNC_rc_repeat

bpf_func_id_BPF_FUNC_redirect

bpf_func_id_BPF_FUNC_redirect_map

bpf_func_id_BPF_FUNC_send_signal

bpf_func_id_BPF_FUNC_set_hash

bpf_func_id_BPF_FUNC_set_hash_invalid

bpf_func_id_BPF_FUNC_setsockopt

bpf_func_id_BPF_FUNC_sk_fullsock

bpf_func_id_BPF_FUNC_sk_lookup_tcp

bpf_func_id_BPF_FUNC_sk_lookup_udp

bpf_func_id_BPF_FUNC_sk_redirect_hash

bpf_func_id_BPF_FUNC_sk_redirect_map

bpf_func_id_BPF_FUNC_sk_release

bpf_func_id_BPF_FUNC_sk_select_reuseport

bpf_func_id_BPF_FUNC_sk_storage_delete

bpf_func_id_BPF_FUNC_sk_storage_get

bpf_func_id_BPF_FUNC_skb_adjust_room

bpf_func_id_BPF_FUNC_skb_ancestor_cgroup_id

bpf_func_id_BPF_FUNC_skb_cgroup_id

bpf_func_id_BPF_FUNC_skb_change_head

bpf_func_id_BPF_FUNC_skb_change_proto

bpf_func_id_BPF_FUNC_skb_change_tail

bpf_func_id_BPF_FUNC_skb_change_type

bpf_func_id_BPF_FUNC_skb_ecn_set_ce

bpf_func_id_BPF_FUNC_skb_get_tunnel_key

bpf_func_id_BPF_FUNC_skb_get_tunnel_opt

bpf_func_id_BPF_FUNC_skb_get_xfrm_state

bpf_func_id_BPF_FUNC_skb_load_bytes

bpf_func_id_BPF_FUNC_skb_load_bytes_relative

bpf_func_id_BPF_FUNC_skb_pull_data

bpf_func_id_BPF_FUNC_skb_set_tunnel_key

bpf_func_id_BPF_FUNC_skb_set_tunnel_opt

bpf_func_id_BPF_FUNC_skb_store_bytes

bpf_func_id_BPF_FUNC_skb_under_cgroup

bpf_func_id_BPF_FUNC_skb_vlan_pop

bpf_func_id_BPF_FUNC_skb_vlan_push

bpf_func_id_BPF_FUNC_skc_lookup_tcp

bpf_func_id_BPF_FUNC_sock_hash_update

bpf_func_id_BPF_FUNC_sock_map_update

bpf_func_id_BPF_FUNC_sock_ops_cb_flags_set

bpf_func_id_BPF_FUNC_spin_lock

bpf_func_id_BPF_FUNC_spin_unlock

bpf_func_id_BPF_FUNC_strtol

bpf_func_id_BPF_FUNC_strtoul

bpf_func_id_BPF_FUNC_sysctl_get_current_value

bpf_func_id_BPF_FUNC_sysctl_get_name

bpf_func_id_BPF_FUNC_sysctl_get_new_value

bpf_func_id_BPF_FUNC_sysctl_set_new_value

bpf_func_id_BPF_FUNC_tail_call

bpf_func_id_BPF_FUNC_tcp_check_syncookie

bpf_func_id_BPF_FUNC_tcp_gen_syncookie

bpf_func_id_BPF_FUNC_tcp_sock

bpf_func_id_BPF_FUNC_trace_printk

bpf_func_id_BPF_FUNC_unspec

bpf_func_id_BPF_FUNC_xdp_adjust_head

bpf_func_id_BPF_FUNC_xdp_adjust_meta

bpf_func_id_BPF_FUNC_xdp_adjust_tail

bpf_func_id___BPF_FUNC_MAX_ID

bpf_hdr_start_off_BPF_HDR_START_MAC

bpf_hdr_start_off_BPF_HDR_START_NET

bpf_lwt_encap_mode_BPF_LWT_ENCAP_IP

bpf_lwt_encap_mode_BPF_LWT_ENCAP_SEG6

bpf_lwt_encap_mode_BPF_LWT_ENCAP_SEG6_INLINE

bpf_map_type_BPF_MAP_TYPE_ARRAY

bpf_map_type_BPF_MAP_TYPE_ARRAY_OF_MAPS

bpf_map_type_BPF_MAP_TYPE_CGROUP_ARRAY

bpf_map_type_BPF_MAP_TYPE_CGROUP_STORAGE

bpf_map_type_BPF_MAP_TYPE_CPUMAP

bpf_map_type_BPF_MAP_TYPE_DEVMAP

bpf_map_type_BPF_MAP_TYPE_DEVMAP_HASH

bpf_map_type_BPF_MAP_TYPE_HASH

bpf_map_type_BPF_MAP_TYPE_HASH_OF_MAPS

bpf_map_type_BPF_MAP_TYPE_LPM_TRIE

bpf_map_type_BPF_MAP_TYPE_LRU_HASH

bpf_map_type_BPF_MAP_TYPE_LRU_PERCPU_HASH

bpf_map_type_BPF_MAP_TYPE_PERCPU_ARRAY

bpf_map_type_BPF_MAP_TYPE_PERCPU_CGROUP_STORAGE

bpf_map_type_BPF_MAP_TYPE_PERCPU_HASH

bpf_map_type_BPF_MAP_TYPE_PERF_EVENT_ARRAY

bpf_map_type_BPF_MAP_TYPE_PROG_ARRAY

bpf_map_type_BPF_MAP_TYPE_QUEUE

bpf_map_type_BPF_MAP_TYPE_REUSEPORT_SOCKARRAY

bpf_map_type_BPF_MAP_TYPE_SK_STORAGE

bpf_map_type_BPF_MAP_TYPE_SOCKHASH

bpf_map_type_BPF_MAP_TYPE_SOCKMAP

bpf_map_type_BPF_MAP_TYPE_STACK

bpf_map_type_BPF_MAP_TYPE_STACK_TRACE

bpf_map_type_BPF_MAP_TYPE_UNSPEC

bpf_map_type_BPF_MAP_TYPE_XSKMAP

bpf_probe_attach_type_BPF_PROBE_ENTRY

bpf_probe_attach_type_BPF_PROBE_RETURN

bpf_prog_type_BPF_PROG_TYPE_CGROUP_DEVICE

bpf_prog_type_BPF_PROG_TYPE_CGROUP_SKB

bpf_prog_type_BPF_PROG_TYPE_CGROUP_SOCK

bpf_prog_type_BPF_PROG_TYPE_CGROUP_SOCKOPT

bpf_prog_type_BPF_PROG_TYPE_CGROUP_SOCK_ADDR

bpf_prog_type_BPF_PROG_TYPE_CGROUP_SYSCTL

bpf_prog_type_BPF_PROG_TYPE_FLOW_DISSECTOR

bpf_prog_type_BPF_PROG_TYPE_KPROBE

bpf_prog_type_BPF_PROG_TYPE_LIRC_MODE2

bpf_prog_type_BPF_PROG_TYPE_LWT_IN

bpf_prog_type_BPF_PROG_TYPE_LWT_OUT

bpf_prog_type_BPF_PROG_TYPE_LWT_SEG6LOCAL

bpf_prog_type_BPF_PROG_TYPE_LWT_XMIT

bpf_prog_type_BPF_PROG_TYPE_PERF_EVENT

bpf_prog_type_BPF_PROG_TYPE_RAW_TRACEPOINT

bpf_prog_type_BPF_PROG_TYPE_RAW_TRACEPOINT_WRITABLE

bpf_prog_type_BPF_PROG_TYPE_SCHED_ACT

bpf_prog_type_BPF_PROG_TYPE_SCHED_CLS

bpf_prog_type_BPF_PROG_TYPE_SK_MSG

bpf_prog_type_BPF_PROG_TYPE_SK_REUSEPORT

bpf_prog_type_BPF_PROG_TYPE_SK_SKB

bpf_prog_type_BPF_PROG_TYPE_SOCKET_FILTER

bpf_prog_type_BPF_PROG_TYPE_SOCK_OPS

bpf_prog_type_BPF_PROG_TYPE_TRACEPOINT

bpf_prog_type_BPF_PROG_TYPE_UNSPEC

bpf_prog_type_BPF_PROG_TYPE_XDP

bpf_ret_code_BPF_DROP

bpf_ret_code_BPF_LWT_REROUTE

bpf_ret_code_BPF_OK

bpf_ret_code_BPF_REDIRECT

bpf_stack_build_id_status_BPF_STACK_BUILD_ID_EMPTY

bpf_stack_build_id_status_BPF_STACK_BUILD_ID_IP

bpf_stack_build_id_status_BPF_STACK_BUILD_ID_VALID

bpf_task_fd_type_BPF_FD_TYPE_KPROBE

bpf_task_fd_type_BPF_FD_TYPE_KRETPROBE

bpf_task_fd_type_BPF_FD_TYPE_RAW_TRACEPOINT

bpf_task_fd_type_BPF_FD_TYPE_TRACEPOINT

bpf_task_fd_type_BPF_FD_TYPE_UPROBE

bpf_task_fd_type_BPF_FD_TYPE_URETPROBE

false_

ifla_geneve_df_GENEVE_DF_INHERIT

ifla_geneve_df_GENEVE_DF_MAX

ifla_geneve_df_GENEVE_DF_SET

ifla_geneve_df_GENEVE_DF_UNSET

ifla_geneve_df___GENEVE_DF_END

ifla_gtp_role_GTP_ROLE_GGSN

ifla_gtp_role_GTP_ROLE_SGSN

ifla_vxlan_df_VXLAN_DF_INHERIT

ifla_vxlan_df_VXLAN_DF_MAX

ifla_vxlan_df_VXLAN_DF_SET

ifla_vxlan_df_VXLAN_DF_UNSET

ifla_vxlan_df___VXLAN_DF_END

in6_addr_gen_mode_IN6_ADDR_GEN_MODE_NONE

in6_addr_gen_mode_IN6_ADDR_GEN_MODE_STABLE_PRIVACY

in6_addr_gen_mode_IN6_ADDR_GEN_MODE_RANDOM

in6_addr_gen_mode_IN6_ADDR_GEN_MODE_EUI64

ipvlan_mode_IPVLAN_MODE_L2

ipvlan_mode_IPVLAN_MODE_L3

ipvlan_mode_IPVLAN_MODE_L3S

ipvlan_mode_IPVLAN_MODE_MAX

macsec_validation_type_MACSEC_VALIDATE_CHECK

macsec_validation_type_MACSEC_VALIDATE_DISABLED

macsec_validation_type_MACSEC_VALIDATE_MAX

macsec_validation_type_MACSEC_VALIDATE_STRICT

macsec_validation_type___MACSEC_VALIDATE_END

macvlan_macaddr_mode_MACVLAN_MACADDR_ADD

macvlan_macaddr_mode_MACVLAN_MACADDR_DEL

macvlan_macaddr_mode_MACVLAN_MACADDR_FLUSH

macvlan_macaddr_mode_MACVLAN_MACADDR_SET

macvlan_mode_MACVLAN_MODE_BRIDGE

macvlan_mode_MACVLAN_MODE_PASSTHRU

macvlan_mode_MACVLAN_MODE_PRIVATE

macvlan_mode_MACVLAN_MODE_SOURCE

macvlan_mode_MACVLAN_MODE_VEPA

nl_mmap_status_NL_MMAP_STATUS_COPY

nl_mmap_status_NL_MMAP_STATUS_RESERVED

nl_mmap_status_NL_MMAP_STATUS_SKIP

nl_mmap_status_NL_MMAP_STATUS_UNUSED

nl_mmap_status_NL_MMAP_STATUS_VALID

nlmsgerr_attrs_NLMSGERR_ATTR_COOKIE

nlmsgerr_attrs_NLMSGERR_ATTR_MAX

nlmsgerr_attrs_NLMSGERR_ATTR_MSG

nlmsgerr_attrs_NLMSGERR_ATTR_OFFS

nlmsgerr_attrs_NLMSGERR_ATTR_UNUSED

nlmsgerr_attrs___NLMSGERR_ATTR_MAX

sk_action_SK_DROP

sk_action_SK_PASS

true_

xdp_action_XDP_ABORTED

xdp_action_XDP_DROP

xdp_action_XDP_PASS

xdp_action_XDP_REDIRECT

xdp_action_XDP_TX

Functions

bcc_create_map⚠

bcc_create_map_xattr⚠

bcc_prog_load⚠

bcc_prog_load_xattr⚠

bpf_attach_kprobe⚠

bpf_attach_perf_event⚠

bpf_attach_perf_event_raw⚠

bpf_attach_raw_tracepoint⚠

bpf_attach_socket⚠

bpf_attach_tracepoint⚠

bpf_attach_uprobe⚠

bpf_attach_xdp⚠

bpf_close_perf_event_fd⚠

bpf_delete_elem⚠

bpf_detach_kprobe⚠

bpf_detach_tracepoint⚠

bpf_detach_uprobe⚠

bpf_get_first_key⚠

bpf_get_next_key⚠

bpf_lookup_elem⚠

bpf_map_get_fd_by_id⚠

bpf_obj_get⚠

bpf_obj_get_info⚠

bpf_obj_pin⚠

bpf_open_perf_buffer⚠

bpf_open_perf_event⚠

bpf_open_raw_sock⚠

bpf_prog_compute_tag⚠

bpf_prog_get_fd_by_id⚠

bpf_prog_get_next_id⚠

bpf_prog_get_tag⚠

bpf_update_elem⚠

pselect⚠

select⚠

Type Definitions

__be16

__be32

__be64

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fd_mask

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__kernel_caddr_t

__kernel_clock_t

__kernel_clockid_t

__kernel_daddr_t

__kernel_gid16_t

__kernel_gid32_t

__kernel_gid_t

__kernel_ino_t

__kernel_ipc_pid_t

__kernel_loff_t

__kernel_long_t

__kernel_mode_t

__kernel_off_t

__kernel_old_dev_t

__kernel_old_gid_t

__kernel_old_uid_t

__kernel_pid_t

__kernel_ptrdiff_t

__kernel_sa_family_t

__kernel_size_t

__kernel_ssize_t

__kernel_suseconds_t

__kernel_time64_t

__kernel_time_t

__kernel_timer_t

__kernel_uid16_t

__kernel_uid32_t

__kernel_uid_t

__kernel_ulong_t

__key_t

__le16

__le32

__le64

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__pthread_list_t

__quad_t

__rlim64_t

__rlim_t

__s8

__s16

__s32

__s64

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u8

__u16

__u32

__u64

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

_bindgen_ty_1

_bindgen_ty_2

_bindgen_ty_3

_bindgen_ty_4

_bindgen_ty_5

_bindgen_ty_6

_bindgen_ty_7

_bindgen_ty_8

_bindgen_ty_9

_bindgen_ty_10

_bindgen_ty_11

_bindgen_ty_12

_bindgen_ty_13

_bindgen_ty_14

_bindgen_ty_15

_bindgen_ty_16

_bindgen_ty_17

_bindgen_ty_18

_bindgen_ty_19

_bindgen_ty_20

_bindgen_ty_21

_bindgen_ty_22

_bindgen_ty_23

_bindgen_ty_24

_bindgen_ty_25

_bindgen_ty_26

_bindgen_ty_27

_bindgen_ty_28

_bindgen_ty_29

_bindgen_ty_30

_bindgen_ty_31

_bindgen_ty_32

_bindgen_ty_33

_bindgen_ty_34

_bindgen_ty_35

_bindgen_ty_36

_bindgen_ty_37

_bindgen_ty_38

_bindgen_ty_39

_bindgen_ty_40

_bindgen_ty_41

_bindgen_ty_42

_bindgen_ty_43

_bindgen_ty_44

_bindgen_ty_45

_bindgen_ty_46

_bindgen_ty_47

blkcnt_t

blksize_t

bpf_adj_room_mode

bpf_attach_type

bpf_cmd

bpf_func_id

bpf_hdr_start_off

bpf_lwt_encap_mode

bpf_map_type

bpf_probe_attach_type

bpf_prog_type

bpf_ret_code

bpf_stack_build_id_status

bpf_task_fd_type

caddr_t

clock_t

clockid_t

daddr_t

dev_t

fd_mask

fsblkcnt_t

fsfilcnt_t

fsid_t

gid_t

id_t

ifla_geneve_df

ifla_gtp_role

ifla_vxlan_df

in6_addr_gen_mode

ino_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

ipvlan_mode

key_t

loff_t

macsec_validation_type

macvlan_macaddr_mode

macvlan_mode

mode_t

nl_mmap_status

nlink_t

nlmsgerr_attrs
enum nlmsgerr_attrs - nlmsgerr attributes @NLMSGERR_ATTR_UNUSED: unused @NLMSGERR_ATTR_MSG: error message string (string) @NLMSGERR_ATTR_OFFS: offset of the invalid attribute in the original message, counting from the beginning of the header (u32) @NLMSGERR_ATTR_COOKIE: arbitrary subsystem specific cookie to be used - in the success case - to identify a created object or operation or similar (binary) @__NLMSGERR_ATTR_MAX: number of attributes @NLMSGERR_ATTR_MAX: highest attribute number

off_t

perf_reader_lost_cb

perf_reader_raw_cb

pid_t

pthread_key_t

pthread_once_t

pthread_spinlock_t

pthread_t

quad_t

register_t

sigset_t

sk_action

suseconds_t

time_t

timer_t

u_char

u_int

u_int8_t

u_int16_t

u_int32_t

u_int64_t

u_long

u_quad_t

u_short

uid_t

uint

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

ulong

ushort

wchar_t

xdp_action

Unions

__kernel_sockaddr_storage__bindgen_ty_1

__pthread_cond_s__bindgen_ty_1

__pthread_cond_s__bindgen_ty_2

__sk_buff__bindgen_ty_1

__sk_buff__bindgen_ty_2

bpf_attr

bpf_attr__bindgen_ty_2__bindgen_ty_1

bpf_attr__bindgen_ty_7__bindgen_ty_1

bpf_fib_lookup__bindgen_ty_1

bpf_fib_lookup__bindgen_ty_2

bpf_fib_lookup__bindgen_ty_3

bpf_flow_keys__bindgen_ty_1

bpf_sock_addr__bindgen_ty_1

bpf_sock_ops__bindgen_ty_1

bpf_sock_ops__bindgen_ty_2

bpf_sock_tuple__bindgen_ty_1

bpf_sockopt__bindgen_ty_1

bpf_sockopt__bindgen_ty_2

bpf_sockopt__bindgen_ty_3

bpf_stack_build_id__bindgen_ty_1

bpf_tunnel_key__bindgen_ty_1

bpf_xfrm_state__bindgen_ty_1

pthread_attr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

sk_msg_md__bindgen_ty_1

sk_msg_md__bindgen_ty_2

sk_reuseport_md__bindgen_ty_1

sk_reuseport_md__bindgen_ty_2